cleanup the ws javax EchoSocket test class
