+++ jetty-websocket/websocket-javax-tests/src/main/java/org/eclipse/jetty/websocket/javax/tests/EventSocket.java
@@ -98,23 +98,4 @@ public class EventSocket
         error = cause;
         errorLatch.countDown();
     }
-
-    @ServerEndpoint("/")
-    @ClientEndpoint
-    public static class EchoSocket extends EventSocket
-    {
-        @Override
-        public void onMessage(String message) throws IOException
-        {
-            super.onMessage(message);
-            session.getBasicRemote().sendText(message);
-        }
-
-        @Override
-        public void onMessage(ByteBuffer message) throws IOException
-        {
-            super.onMessage(message);
-            session.getBasicRemote().sendBinary(message);
-        }
-    }
 }
